Bug 1111190 - Fix sign-compare warnings in CollectIonReports; r=froydnj We only use kNumFields to compare it to the result of int sscanf(...).
UltraBlame original commit: d13ba6a6b8b97a237f846e89fac4731897d57b19

+++ xpcom/base/SystemMemoryReporter.cpp
@@ -718,7 +718,7 @@ private:
     
     
     const char* const kFormatString = "%63s %" SCNu32 " %" SCNu64;
-    const size_t kNumFields = 3;
+    const int kNumFields = 3;
     const size_t kStringSize = 64;
     const char* const kIonIommuPath = "/sys/kernel/debug/ion/iommu";
